requirements.txt
annotated-types          0.7.0
anyio                    4.7.0
certifi                  2024.12.14
charset-normalizer       3.4.0
contourpy                1.3.1
cycler                   0.12.1
distro                   1.9.0
filelock                 3.16.1
fonttools                4.55.3
fsspec                   2024.10.0
groq                     0.13.1
h11                      0.14.0
httpcore                 1.0.7
httpx                    0.28.1
idna                     3.10
Jinja2                   3.1.4
jiter                    0.8.2
kiwisolver               1.4.7
MarkupSafe               3.0.2
matplotlib               3.10.0
mpmath                   1.3.0
networkx                 3.4.2
numpy                    2.2.0
nvidia-cublas-cu12       12.4.5.8
nvidia-cuda-cupti-cu12   12.4.127
nvidia-cuda-nvrtc-cu12   12.4.127
nvidia-cuda-runtime-cu12 12.4.127
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.2.1.3
nvidia-curand-cu12       10.3.5.147
nvidia-cusolver-cu12     11.6.1.9
nvidia-cusparse-cu12     12.3.1.170
nvidia-nccl-cu12         2.21.5
nvidia-nvjitlink-cu12    12.4.127
nvidia-nvtx-cu12         12.4.127
openai                   1.57.4
opencv-python            4.10.0.84
packaging                24.2
pandas                   2.2.3
pillow                   11.0.0
pip                      24.3.1
psutil                   6.1.0
py-cpuinfo               9.0.0
pydantic                 2.10.3
pydantic_core            2.27.1
pyparsing                3.2.0
python-dateutil          2.9.0.post0
pyttsx3                  2.98
pytz                     2024.2
PyYAML                   6.0.2
requests                 2.32.3
scipy                    1.14.1
seaborn                  0.13.2
setuptools               75.6.0
six                      1.17.0
sniffio                  1.3.1
sympy                    1.13.1
torch                    2.5.1
torchvision              0.20.1
tqdm                     4.67.1
triton                   3.1.0
typing_extensions        4.12.2
tzdata                   2024.2
ultralytics              8.3.49
ultralytics-thop         2.0.13
urllib3                  2.2.3